+++ src/pages/requests/ApplicationProcess/interface.tsx
@@ -1,27 +1,15 @@
 import { useParams, useLocation } from "react-router-dom";
-import { Stack, Text, useMediaQuery, Button, Select } from "@inubekit/inubekit";
 import { useState, useEffect } from "react";
 
-import { AppMenu } from "@components/layout/AppMenu";
 import { IRoute } from "@components/layout/AppMenu/types";
-import { spacing } from "@design/tokens/spacing";
 import { requestConfigs } from "@config/requests.config";
 import { capitalizeFullName } from "@utils/string";
 import { useHumanResourceRequest } from "@hooks/useHumanResourceRequestById";
 import { useEvaluateResponsibleOfTasks } from "@hooks/useEvaluateResponsibleOfTasks";
 import { useHeaders } from "@hooks/useHeaders";
 import { useHumanDecisionTasks } from "@hooks/useHumanDecisionTasks";
-import {
-  HumanDecision,
-  HumanDecisionTranslations,
-} from "@ptypes/humanResources.types";
 
-import { StyledFieldsetContainer } from "./styles";
-import { RequestSummary } from "./Components/RequestSummary";
-import { ActionModal } from "./Components/Actions";
-import { Fieldset } from "@components/data/Fieldset";
-
-interface ApplicationProcessUIProps {
+export interface ApplicationProcessUIProps {
   appName: string;
   appRoute: IRoute[];
   navigatePage: string;
@@ -35,8 +23,7 @@ function isRequestConfigKey(
   return value in requestConfigs;
 }
 
-function ApplicationProcessUI(props: ApplicationProcessUIProps) {
-  const { appRoute, navigatePage } = props;
+export function useApplicationProcessLogic(appRoute: IRoute[]) {
   const { id } = useParams<{ id: string }>();
   const { state } = useLocation() as {
     state?: {
@@ -48,7 +35,6 @@ function ApplicationProcessUI(props: ApplicationProcessUIProps) {
     };
   };
 
-  const isMobile = useMediaQuery("(max-width: 1000px)");
   const [showActions, setShowActions] = useState(false);
   const [decision, setDecision] = useState<string>("");
 
@@ -135,103 +121,28 @@ function ApplicationProcessUI(props: ApplicationProcessUIProps) {
   const handleSeeAttachments = () => console.log("Ver adjuntos");
   const handleSend = () => console.log("Decisión seleccionada:", decision);
 
-  return (
-    <AppMenu
-      appName={displayRequestLabel}
-      appRoute={updatedAppRoute}
-      navigatePage={navigatePage}
-      appDescription={displayDescription}
-    >
-      <Stack direction="column" gap={spacing.s200}>
-        {isMobile && showActions && (
-          <ActionModal
-            onExecute={handleExecute}
-            onDiscard={handleDiscard}
-            onAttach={handleAttach}
-            onSeeAttachments={handleSeeAttachments}
-            onClose={() => setShowActions(false)}
-            actionDescriptions={{
-              execute: "No puedes ejecutar esta acción ahora",
-              discard: "No puedes descartar esta acción ahora",
-              attach: "No puedes adjuntar archivos en este momento",
-              seeAttachments: "No puedes ver los adjuntos en este momento",
-            }}
-          />
-        )}
-
-        <RequestSummary
-          requestNumber={
-            requestData?.humanResourceRequestNumber ??
-            state?.requestNumber ??
-            id
-          }
-          requestDate={
-            requestData?.humanResourceRequestDate ?? state?.requestDate
-          }
-          title={finalRequestLabel}
-          status={requestData?.humanResourceRequestStatus ?? state?.status}
-          fullStaffName={capitalizeFullName(
-            state?.fullStaffName ?? "Sin responsable",
-          )}
-          humanResourceRequestData={requestData?.humanResourceRequestData}
-          requestType={requestData?.humanResourceRequestType}
-          isLoading={isLoadingRequest}
-        />
-
-        <StyledFieldsetContainer $isMobile={isMobile}>
-          <Fieldset
-            title="Por hacer"
-            descriptionTitle={
-              loading
-                ? "Cargando..."
-                : error
-                  ? "Error al cargar"
-                  : responsibleLabel
-            }
-          >
-            <Stack direction="column" gap={spacing.s150}>
-              <Text>Verificar viabilidad de la solicitud.</Text>
-
-              <Stack direction="row" alignItems="flex-end" gap={spacing.s150}>
-                <Select
-                  name="decision"
-                  id="decision"
-                  label="Decisión"
-                  placeholder={
-                    loadingDecisions
-                      ? "Cargando opciones..."
-                      : errorDecisions
-                        ? "Error al cargar"
-                        : "Seleccione una opción"
-                  }
-                  options={
-                    decisionsData?.decisions.map((opt) => ({
-                      id: opt,
-                      value: opt,
-                      label:
-                        HumanDecisionTranslations[opt as HumanDecision] ?? opt,
-                    })) ?? []
-                  }
-                  value={decision}
-                  onChange={(_, value) => setDecision(value)}
-                  size="wide"
-                  fullwidth
-                />
-
-                <Button
-                  appearance="primary"
-                  variant="filled"
-                  onClick={handleSend}
-                >
-                  Enviar
-                </Button>
-              </Stack>
-            </Stack>
-          </Fieldset>
-        </StyledFieldsetContainer>
-      </Stack>
-    </AppMenu>
-  );
+  return {
+    id,
+    state,
+    decision,
+    setDecision,
+    showActions,
+    setShowActions,
+    requestData,
+    isLoadingRequest,
+    responsibleLabel,
+    loading,
+    error,
+    loadingDecisions,
+    errorDecisions,
+    decisionsData,
+    updatedAppRoute,
+    displayRequestLabel,
+    displayDescription,
+    handleDiscard,
+    handleExecute,
+    handleAttach,
+    handleSeeAttachments,
+    handleSend,
+  };
 }
-
-export { ApplicationProcessUI };
